+++ src/Element/HasHref.php
@@ -26,7 +26,7 @@ use UnitEnum;
  * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/a#href
  * @link https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/href
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 trait HasHref