+++ src/Global/CanBeHidden.php
@@ -25,7 +25,7 @@ use UnitEnum;
  *
  * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/hidden
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 trait CanBeHidden