+++ src/Global/HasMicroData.php
@@ -30,7 +30,7 @@ use UnitEnum;
  * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/itemscope
  * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/itemtype
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 trait HasMicroData