+++ src/Global/HasSpellcheck.php
@@ -29,7 +29,7 @@ use function is_bool;
  *
  * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/spellcheck
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 trait HasSpellcheck