+++ src/Values/ElementAttribute.php
@@ -16,6 +16,9 @@ namespace UIAwesome\Html\Attribute\Values;
  * - Integration-ready for tag rendering and element generation APIs.
  *
  * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Element
+ *
+ * @copyright Copyright (C) 2026 Terabytesoftw.
+ * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 enum ElementAttribute: string
 {