+++ src/Values/Language.php
@@ -18,7 +18,7 @@ namespace UIAwesome\Html\Attribute\Values;
  * @link https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/lang
  * @link https://tools.ietf.org/html/bcp47
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 enum Language: string