+++ tests/Global/HasDraggableTest.php
@@ -28,7 +28,7 @@ use UnitEnum;
  *
  * {@see DraggableProvider} for test case data providers.
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 #[Group('global')]