+++ tests/HasCrossoriginTest.php
@@ -28,7 +28,7 @@ use UnitEnum;
  *
  * {@see CrossoriginProvider} for test case data providers.
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 #[Group('attribute')]
@@ -91,7 +91,7 @@ final class HasCrossoriginTest extends TestCase
         );
     }
 
-    public function testThrowExceptionWhenSettingInvalidCrossoriginValue(): void
+    public function testThrowInvalidArgumentExceptionWhenSettingCrossoriginValue(): void
     {
         $instance = new class {
             use HasAttributes;
@@ -103,7 +103,7 @@ final class HasCrossoriginTest extends TestCase
             Message::VALUE_NOT_IN_LIST->getMessage(
                 'invalid-value',
                 Attribute::CROSSORIGIN->value,
-                implode('\', \'', Enum::normalizeArray(Crossorigin::cases())),
+                implode("', '", Enum::normalizeArray(Crossorigin::cases())),
             ),
         );