+++ tests/Support/Provider/CrossoriginProvider.php
@@ -13,7 +13,7 @@ use UnitEnum;
  *
  * Provides representative input/output pairs for the `crossorigin` attribute.
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 final class CrossoriginProvider