+++ tests/Support/Provider/Global/AriaProvider.php
@@ -14,7 +14,7 @@ use UnitEnum;
  *
  * Provides representative input/output pairs for `aria-*` attributes.
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 final class AriaProvider